requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiohttp-retry==2.9.1
aiosignal==1.4.0
aiosqlite==0.21.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
arrow==1.4.0
attrs==25.4.0
babel==2.18.0
beautifulsoup4==4.14.3
blinker==1.9.0
cartesia==3.0.2
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
colorlog==6.10.1
comtypes==1.4.16
cryptography==46.0.3
csvw==3.7.0
dataclasses-json==0.6.7
dateparser==1.2.2
DateTime==6.0
deepgram-sdk==6.0.1
distro==1.9.0
dlinfo==2.0.0
docopt==0.6.2
edge-tts==7.2.7
elevenlabs==2.39.0
espeakng-loader==0.2.4
faiss-cpu==1.13.1
fastapi==0.128.0
ffmpeg-python==0.2.0
filelock==3.20.0
Flask==3.1.2
flatbuffers==25.12.19
frozenlist==1.8.0
fsspec==2025.12.0
future==1.0.0
greenlet==3.3.0
groq==0.37.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
huggingface-hub==0.36.0
ics==0.7.2
idna==3.11
ImageIO==2.37.3
imageio-ffmpeg==0.6.0
isodate==0.7.2
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.13.0
joblib==1.5.2
Js2Py==0.74
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kokoro-onnx==0.4.7
langchain==1.1.2
langchain-classic==1.0.0
langchain-community==0.4.1
langchain-core==1.2.19
langchain-groq==1.1.0
langchain-huggingface==1.1.0
langchain-openai==1.1.11
langchain-text-splitters==1.0.0
langgraph==1.0.4
langgraph-checkpoint==3.0.1
langgraph-checkpoint-sqlite==3.0.0
langgraph-prebuilt==1.0.5
langgraph-sdk==0.2.14
langsmith==0.4.56
language-tags==1.2.0
llvmlite==0.46.0
MarkupSafe==3.0.3
marshmallow==3.26.1
more-itertools==10.8.0
mpmath==1.3.0
multidict==6.7.0
mypy_extensions==1.1.0
networkx==3.6
numba==0.64.0
numpy==2.3.5
onnxruntime==1.24.4
openai==2.28.0
openai-whisper==20250625
orjson==3.11.5
ormsgpack==1.12.0
packaging==25.0
phonemizer-fork==3.3.1
pillow==12.0.0
pipwin==0.5.2
propcache==0.4.1
protobuf==7.34.0
psutil==7.2.2
pycparser==2.23
pycryptodome==3.23.0
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
pydub==0.25.1
pyjsparser==2.7.1
PyJWT==2.12.1
pyparsing==3.3.2
pypiwin32==223
PyPrind==2.11.3
pySmartDL==1.3.4
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pyttsx3==2.99
pytz==2025.2
pywin32==311
PyYAML==6.0.3
rdflib==7.6.0
redis==7.1.0
referencing==0.37.0
regex==2025.11.3
requests==2.32.5
requests-toolbelt==1.0.0
rfc3986==1.5.0
rpds-py==0.30.0
safetensors==0.7.0
sarvamai==0.1.27
scikit-learn==1.7.2
scipy==1.16.3
segments==2.4.0
sentence-transformers==5.1.2
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
sounddevice==0.5.5
soundfile==0.13.1
soupsieve==2.8.3
SQLAlchemy==2.0.44
sqlite-vec==0.1.6
starlette==0.50.0
sympy==1.14.0
tabulate==0.10.0
TatSu==5.13.2
tenacity==9.1.2
termcolor==3.3.0
threadpoolctl==3.6.0
tiktoken==0.12.0
tokenizers==0.22.1
torch==2.9.1
tqdm==4.67.1
transformers==4.57.3
twilio==9.10.3
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.6.0
uuid_utils==0.12.0
uvicorn==0.42.0
websockets==16.0
Werkzeug==3.1.4
xxhash==3.6.0
yarl==1.22.0
zope.interface==8.1.1
zstandard==0.25.0
